refactored imgur upload code into multiple files: upload - storageEngine - client
